requirements.txt
flask
flask-socketio
google-cloud-vision
eventlet